Refactor PanelExpansion to its own separate file
