+++ src/components/index.ts
@@ -22,9 +22,5 @@ export {
   ToggleInput,
   ToggleInputProps,
 } from './Form/Toggle';
-export {
-  Panel,
-  PanelProps,
-  PanelExpansion,
-  PanelExpansionProps,
-} from './Panel/Panel';
+export { Panel, PanelProps } from './Panel/Panel';
+export { PanelExpansion, PanelExpansionProps } from './Panel/PanelExpansion';
